css
css

+++ main/views.py
@@ -224,8 +224,9 @@ def editpost(request,pk):
     return render(request, 'main/editpost.html',context = mydict)
 
 def deletepost(request,pk):
-    fuser = Post.objects.get(id=pk)
-    pimage = PostImages.objects.filter(gallery=fuser)
-    fuser.delete()
-    pimage.delete()
+    if request.method == 'POST':
+        fuser = Post.objects.get(id=pk)
+        pimage = PostImages.objects.filter(gallery=fuser)
+        fuser.delete()
+        pimage.delete()
     return redirect('listaddedposts')
